pyproject.toml
[tool.poetry]
name = "gsmtasks-complete"
version = "0.1.0"
description = ""
authors = ["Raphael Krupinski <10319569-mattesilver@users.noreply.gitlab.com>"]
readme = "README.md"
packages = [{include = "gsmtasks_complete", from = "src"}]

[tool.poetry.dependencies]
gsmtasks-client = {git = "https://github.com/oeklo/gsmtasks-client.git", rev = "develop"}
python = "^3.9"
asyncclick = "^8.1.3.4"
#gsmtasks-client = {path='../gsmtasks/gsmtasks-client', develop=true}
pytz = "^2022.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
